+++ models/index.js
@@ -1,5 +1,5 @@
 module.exports = {
-  Book: require("./book"),
+  User: require("./user"),
   Item: require("./item"),
   Stats: require("./stats")
 };
