+++ routes/api/index.js
@@ -1,7 +1,11 @@
 const router = require("express").Router();
-const bookRoutes = require("./books");
+const itemsRoutes = require("./books");
+const statsRoutes = require("./stats");
+const userRoutes = require("./user");
 
 // Book routes
-router.use("/books", bookRoutes);
+router.use("/items", itemsRoutes);
+router.use("/stats", statsRoutes);
+router.use("/user", userRoutes);
 
 module.exports = router;
