+++ routes/api/user.js
@@ -3,7 +3,7 @@ const FFController = require("../../controllers/FFController");
 
 // Matches with "/api/books"
 router.route("/") //find out where it will be connecting to front end js
-    .get(FFController.findAllItem)
+    .get(FFController.findAllUser)
     .post(FFController.create);
 
 // Matches with "/api/books/:id"
